requirements.txt
requests~=2.32.4
Jinja2~=3.1.6
uiautomator2~=3.3.3
wda~=0.3.0
facebook-wda~=1.5.0
allure-python-commons~=2.14.3
pillow~=10.4.0
pytest~=8.4.1